refactor(digest): share scheduler lifecycle
